pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "coverage>=7.9.2",
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
]
